Agregar BigInteger

+++ src/main/java/com/sistema/examenes/services/Competencia_ServiceImpl.java
@@ -8,6 +8,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Service;
 
+import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -38,7 +39,7 @@ public class Competencia_ServiceImpl extends GenericServiceImpl<Competencia, Lon
 
         for (Object[] resultado : resultados) {
             Competencia_DTO competenciaDTO = new Competencia_DTO(
-                    (Long) resultado[0],
+                    ((BigInteger) resultado[0]).longValue(),
                     (String) resultado[1]
             );
             competenciasEncontradas.add(competenciaDTO);
