+++ src/main/java/com/sistema/examenes/services/Eje_ServiceImpl.java
@@ -9,6 +9,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Service;
 
+import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -34,7 +35,7 @@ public class Eje_ServiceImpl extends GenericServiceImpl<Eje, Long> implements Ej
 
         for (Object[] resultado : resultados) {
             Eje_DTO ejeDTO = new Eje_DTO(
-                    (Long) resultado[0],
+                    ((BigInteger) resultado[0]).longValue(),
                     (String) resultado[1]
             );
             ejesEncontrados.add(ejeDTO);
