+++ src/main/java/com/sistema/examenes/services/ObjetivoPND_ServiceImpl.java
@@ -8,6 +8,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Service;
 
+import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -33,7 +34,7 @@ public class ObjetivoPND_ServiceImpl extends GenericServiceImpl<ObjetivoPND, Lon
 
         for (Object[] resultado : resultados) {
             ObjetivoPnd_DTO objetivoDTO = new ObjetivoPnd_DTO(
-                    (Long) resultado[0],
+                    ((BigInteger) resultado[0]).longValue(),
                     (String) resultado[1],
                     (String) resultado[2],
                     (Long) resultado[3],
