+++ src/main/java/com/sistema/examenes/services/ObjetivoPdot_ServiceImpl.java
@@ -8,6 +8,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Service;
 
+import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -39,7 +40,7 @@ public class ObjetivoPdot_ServiceImpl extends GenericServiceImpl<Objetivo_pdot, 
 
         for (Object[] resultado : resultados) {
             ObjetivoPdot_DTO objetivoDTO = new ObjetivoPdot_DTO(
-                    (Long) resultado[0],
+                    ((BigInteger) resultado[0]).longValue(),
                     (String) resultado[1],
                     (String) resultado[2]
             );
